car-recsys-system/backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Vector Database
qdrant-client==1.7.0

# LangChain & OpenAI (for chatbot)
langchain==0.2.16
langchain-openai==0.1.23
langchain-core==0.2.38
openai==1.42.0
tiktoken==0.7.0

# Cache
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Utilities
numpy==1.26.3
scikit-learn==1.4.0
httpx==0.26.0
python-dotenv==1.0.0
